// enums/navigationMenu/yourAssessmentSubMenu.ts
export enum YourAssessmentSubMenu {
  PropertyAssessmentAndTaxes = 'Property Assessment and Property Taxes',
  NoticesAndNotifications = 'Notices and Notifications',
  ReadingAssessmentNotice = 'Reading your property assessment notice',
  HomeownersHub = "Homeowners’ Hub",
  HowSalesAffectAssessment = 'How sales affect your assessment',
  PropertyInspections = 'Property Inspections',
  ThreeApproachesToValue = 'Three approaches to value',
  FourYearAssessmentCycle = 'The four year assessment cycle',
  ComparePropertyValues = 'Compare property values',
  Disclosure2016Update = 'Disclosure for the 2016 property assessment update',
}